+++ upload-cli/.env.example
@@ -1,10 +1,10 @@
-BUCKET_NAME=
+export BUCKET_NAME=
 
-BUCKET_NAME_DEV=
-GOOGLE_CREDENTIALS_DEV=
+export BUCKET_NAME_DEV=
+export GOOGLE_CREDENTIALS_DEV=
 
-BUCKET_NAME_PROD=
-GOOGLE_CREDENTIALS_PROD=
+export BUCKET_NAME_PROD=
+export GOOGLE_CREDENTIALS_PROD=
 
 export GOOGLE_CLOUD_KEYFILE_JSON=
 export GOOGLE_APPLICATION_CREDENTIALS="$GOOGLE_CLOUD_KEYFILE_JSON"
