+++ upload-cli/package.json
@@ -7,7 +7,6 @@
   },
   "dependencies": {
     "@google-cloud/storage": "^4.5.0",
-    "@types/sharp": "^0.24.0",
     "chalk": "^3.0.0",
     "dotenv": "^8.2.0",
     "fluent-ffmpeg": "^2.1.2",
@@ -18,6 +17,7 @@
   },
   "devDependencies": {
     "@types/fluent-ffmpeg": "^2.1.14",
-    "@types/inquirer": "^6.5.0"
+    "@types/inquirer": "^6.5.0",
+    "@types/sharp": "^0.24.0"
   }
 }
