+++ upload-cli/src/lib/uploader/index.ts
@@ -1,7 +1,5 @@
 import { promises as fs } from 'fs';
 import path from 'path';
-import dotenv from 'dotenv';
-dotenv.config();
 
 export interface OriginalTime {
   year: number;
@@ -24,15 +22,6 @@ export interface UploadFilePath extends OriginalTime {
   fileExtension?: string;
 }
 
-const {
-  BUCKET_NAME_DEV,
-  BUCKET_NAME_PROD,
-  GOOGLE_CREDENTIALS_DEV,
-  GOOGLE_CREDENTIALS_PROD,
-} = process.env;
-
-export let BUCKET_NAME: string;
-
 export const DELAY_TIME = 100;
 
 export const getFilesByPath = async (path: string): Promise<string[]> => {
@@ -43,28 +32,6 @@ export const getFilesByPath = async (path: string): Promise<string[]> => {
     throw Error(`Unable to get files from ${path}. ${e}`);
   }
 };
-
-export const setEnvironment = async (
-  target: UploaderIO['target'],
-): Promise<void> =>
-  new Promise(resolve => {
-    if (target === 'development') {
-      BUCKET_NAME = BUCKET_NAME_DEV ?? '';
-      process.env = {
-        ...process.env,
-        GOOGLE_APPLICATION_CREDENTIALS: GOOGLE_CREDENTIALS_DEV,
-      };
-    }
-
-    if (target === 'production') {
-      BUCKET_NAME = BUCKET_NAME_PROD ?? '';
-      process.env = {
-        ...process.env,
-        GOOGLE_APPLICATION_CREDENTIALS: GOOGLE_CREDENTIALS_PROD,
-      };
-    }
-    resolve();
-  });
 
 export const getFilePath = ({
   source,
